+++ lib/models/models.dart
@@ -7,3 +7,5 @@ export 'notes/notebook_model.dart';
 export 'mail/mail_models.dart';
 // Export group models
 export 'group/group_models.dart';
+export 'planner/planner_models.dart';
+export 'todo/todo_models.dart';